dom/system/tests/ioutils/test_ioutils_mkdir.html

<!-- Any copyright is dedicated to the Public Domain.
- http://creativecommons.org/publicdomain/zero/1.0/ -->
<!DOCTYPE HTML>
<html>

<head>
  <meta charset="utf-8">
  <title>Test the IOUtils file I/O API</title>
  <script src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" href="chrome://mochikit/content/tests/SimpleTest/test.css" />
  <script src="file_ioutils_test_fixtures.js"></script>
  <script>
    "use strict";

    const { Assert } = ChromeUtils.import("resource://testing-common/Assert.jsm");
    const { ObjectUtils } = ChromeUtils.import("resource://gre/modules/ObjectUtils.jsm");

    // TODO: Remove this import for OS.File. It is currently being used as a
    //       stop gap for missing IOUtils functionality.
    const { OS } = ChromeUtils.import("resource://gre/modules/osfile.jsm");


    const tmpDir = OS.Constants.Path.tmpDir;

    add_task(async function test_make_directory() {
      info("Test creating a new directory");
      const newDirectoryName = OS.Path.join(tmpDir, "test_ioutils_new_dir.tmp.d");
      await IOUtils.makeDirectory(newDirectoryName);
      ok(
        await OS.File.exists(newDirectoryName),
        "IOUtils::makeDirectory can create a new directory"
      );

      info("Test creating an existing directory");
      await IOUtils.makeDirectory(newDirectoryName, { ignoreExisting: true });
      ok(
        await OS.File.exists(newDirectoryName),
        "IOUtils::makeDirectory can ignore existing directories"
      );
      await Assert.rejects(
        IOUtils.makeDirectory(newDirectoryName, { ignoreExisting: false }),
        /Could not create directory because it already exists at .*/,
        "IOUtils::makeDirectory can throw if the target dir exists"
      )

      info("Test creating a nested directory");
      const parentDirName = OS.Path.join(tmpDir, "test_ioutils_mkdir_parent.tmp.d");
      const nestedDirName = OS.Path.join(
        parentDirName,
        "test_ioutils_mkdir_child.tmp.d"
      );
      await Assert.rejects(
        IOUtils.makeDirectory(nestedDirName, { createAncestors: false }),
        /Could not create directory at .* because the path has missing ancestor components/,
        "IOUtils::makeDirectory can fail if the target is missing parents"
      );
      ok(!await OS.File.exists(nestedDirName), `Expected ${nestedDirName} not to exist`);
      await IOUtils.makeDirectory(nestedDirName, { createAncestors: true });
      ok(
        await OS.File.exists(nestedDirName),
        "IOUtils::makeDirectory can create ancestors of the target directory"
      );

      await cleanup(newDirectoryName, parentDirName);
    });

    add_task(async function test_make_directory_failure() {
      info("Try to create a directory where a file already exists");
      const notADirFileName = OS.Path.join(tmpDir, "test_ioutils_not_a_dir.tmp");
      await createFile(notADirFileName);

      await Assert.rejects(
        IOUtils.makeDirectory(notADirFileName, { ignoreExisting: false }),
        /Could not create directory because the target file\(.*\) exists and is not a directory/,
        "IOUtils::makeDirectory [ignoreExisting: false] throws when the target is an existing file"
      );
      ok(await fileExists(notADirFileName), `Expected ${notADirFileName} to exist`);

      await Assert.rejects(
        IOUtils.makeDirectory(notADirFileName, { ignoreExisting: true }),
        /Could not create directory because the target file\(.*\) exists and is not a directory/,
        "IOUtils::makeDirectory [ignoreExisting: true] throws when the target is an existing file"
      );
      ok(await fileExists(notADirFileName), `Expected ${notADirFileName} to exist`);

      await cleanup(notADirFileName);
    });
  </script>
</head>

<body>
  <p id="display"></p>
  <div id="content" style="display: none"></div>
  <pre id="test"></pre>
</body>

</html>
